requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
ecdsa==0.19.1
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
idna==3.10
passlib==1.7.4
pillow==11.1.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.0
pydantic_core==2.33.0
python-jose==3.4.0
python-multipart==0.0.20
qrcode==8.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.0
uvicorn==0.34.0
pydantic-settings==2.9.1 
psycopg2-binary
pydantic[email]
alembic
python-dotenv
loguru
pytest
httpx